done 2

+++ config/jwt.php
@@ -4,10 +4,9 @@ return [
 
     'secret' => env('JWT_SECRET'),
 
-
-    'ttl' => env('JWT_TTL', 60),
-
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 1440),
+    'ttl' => (int) env('JWT_TTL', 60),
+    
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 1440),
 
     'algo' => env('JWT_ALGO', 'HS256'),
